final-pjt-back/finlife/serializers.py
from rest_framework import serializers
from .models import DepositProducts, DepositOptions
# , DepositOptions

class DepositProductsSerializer(serializers.ModelSerializer):
    
    class Meta:
        model = DepositProducts
        fields = '__all__'

class DepositOptionsSerializer(serializers.ModelSerializer):
    class Meta:
        model = DepositOptions
        fields = '__all__'
        read_only_fields = ('fin_prdt_cd',)